# Optimize OP_BRANCH - if possible check the first char before enter recursion