In theory patterns without branches (like `/./`) could be a tick slower. Yet in reality, in the benchmark they are not.

However with the current parser, this - as it is now - is already the edge of tricky. With a rewritten parser, there could be a few similar optimizations for certain loops too.

Anyway the timings speak for themself. 

```
Running 5 repeats for each benchmark. Use -c <n> to change

                                                        Time   before |    after  | Match count
==============================================================================
                                                  /Twain/ :     31 ms |     40 ms |        2388
                                              /(?i)Twain/ :     59 ms |     62 ms |        2657
                                             /[a-z]shing/ :    228 ms |    228 ms |        1877
                             /Huck[a-zA-Z]+|Saw[a-zA-Z]+/ :     31 ms |     40 ms |         396
                                                      /./ :    418 ms |    421 ms |    18905427
                                                    /(.)/ :    662 ms |    662 ms |    18905427
                                                      /e/ :     78 ms |     84 ms |     1781425
                                                    /(e)/ :    100 ms |    109 ms |     1781425
                                           /(?s).{1,45} / :     37 ms |     37 ms |      475715
                                         /(?s)\G.{1,45} / :    140 ms |    146 ms |       10616
                                         /\G(?s).{1,45} / :     15 ms |     12 ms |       10616
                                          /(?s).{1,45}? / :    159 ms |    153 ms |     3241534
                                        /(?s)\G.{1,45}? / :    143 ms |    150 ms |       69431
                                        /\G(?s).{1,45}? / :     18 ms |     18 ms |       69431
                                              /\b\w+nn\b/ :    221 ms |    243 ms |         359
                                       /[a-q][^u-z]{13}x/ :    618 ms |    628 ms |        4929
                            /Tom|Sawyer|Huckleberry|Finn/ :     34 ms |     43 ms |        2757
                        /(?i)Tom|Sawyer|Huckleberry|Finn/ :    165 ms |    125 ms |        4538
                    /.{0,2}(Tom|Sawyer|Huckleberry|Finn)/ :   2131 ms |   1487 ms |        2757    !!!!!!!
                    /.{2,4}(Tom|Sawyer|Huckleberry|Finn)/ :   2353 ms |   1696 ms |        2028    !!!!!!!
                      /Tom.{10,25}river|river.{10,25}Tom/ :     53 ms |     53 ms |           2
                                           /[a-zA-Z]+ing/ :    584 ms |    575 ms |       95863
                                  /\s[a-zA-Z]{0,12}ing\s/ :    268 ms |    259 ms |       67810
                          /([A-Za-z]awyer|[A-Za-z]inn)\s/ :    525 ms |    543 ms |         313
                              /["'][^"']{0,30}[?!\.]["']/ :     68 ms |     78 ms |        9857
                                /Tom(.{3,3}|.{5,5})*Finn/ :   2884 ms |   2740 ms |          11
                                    /Tom(...|.....)*Finn/ :   1240 ms |   1590 ms |          11    !!!!!!! Bad
                                   /Tom(...|.....)*?Finn/ :   1353 ms |   1521 ms |          11    !!!!!!! Bad
                      /Tom((...|.....){2,9}\s){1,5}?Finn/ :   3797 ms |   3628 ms |          11
                      /Tom((...|.....){2,9}?\s){1,5}Finn/ :   3634 ms |   3631 ms |          11
                                        /\G(?is).(?=.*$)/ :    915 ms |    925 ms |    20045118
                                /\G(?is).(?=(.){1,5}?$)?/ :   3362 ms |   3428 ms |    20045118
                                       /\G(?is).(?=.*+$)/ :    897 ms |    900 ms |    20045118
                /\G(?is).{10,10}(?=(e|y|on|fin|.){0,20})/ :   1881 ms |    171 ms |     2004511    !!!!!!!
              /\G(?is).{10,10}(?=(?>e|y|on|fin|.){0,20})/ :   1871 ms |    171 ms |     2004511    !!!!!!!
                                          /Tom(?!.*Finn)/ :     31 ms |     40 ms |        2441
     /(?i)(?>[aeoui][bcdfghjklmnpqrstvwxyz \r\n]*){1,40}/ :    625 ms |    568 ms |      579843
                          /(?i)[ bdlm][abdegij][mnoprst]/ :    225 ms |    218 ms |      788955
Total:                                                       31868 ms |  27440 ms |
```

A real live app of mine gained 40% :)